[AS7-6608] Initial removal of JAXRPC support
