--- webservices/server-integration/src/main/java/org/jboss/as/webservices/tomcat/SecurityMetaDataAccessorEJB21.java
@@ -1,44 +1,0 @@
-/*
- * JBoss, Home of Professional Open Source.
- * Copyright 2009, Red Hat Middleware LLC, and individual contributors
- * as indicated by the @author tags. See the copyright.txt file in the
- * distribution for a full listing of individual contributors.
- *
- * This is free software; you can redistribute it and/or modify it
- * under the terms of the GNU Lesser General Public License as
- * published by the Free Software Foundation; either version 2.1 of
- * the License, or (at your option) any later version.
- *
- * This software is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public
- * License along with this software; if not, write to the Free
- * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
- * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
- */
-package org.jboss.as.webservices.tomcat;
-
-import java.util.List;
-
-import org.jboss.as.webservices.metadata.model.EJBEndpoint;
-import org.jboss.as.webservices.metadata.model.JAXRPCDeployment;
-import org.jboss.ws.common.integration.WSHelper;
-import org.jboss.wsf.spi.deployment.Deployment;
-
-/**
- * Creates web app security meta data for EJB 21 deployment.
- *
- * @author <a href="mailto:ropalka@redhat.com">Richard Opalka</a>
- */
-final class SecurityMetaDataAccessorEJB21 extends AbstractSecurityMetaDataAccessorEJB {
-
-    @Override
-    protected List<EJBEndpoint> getEjbEndpoints(final Deployment dep) {
-        final JAXRPCDeployment jaxrpcDeployment = WSHelper.getRequiredAttachment(dep, JAXRPCDeployment.class);
-        return jaxrpcDeployment.getEjbEndpoints();
-    }
-
-}
