+++ webservices/server-integration/src/main/java/org/jboss/as/webservices/util/ASHelper.java
@@ -22,7 +22,6 @@
 package org.jboss.as.webservices.util;
 
 import static org.jboss.as.webservices.util.DotNames.JAXWS_SERVICE_CLASS;
-import static org.jboss.as.webservices.util.WSAttachmentKeys.JAXRPC_ENDPOINTS_KEY;
 import static org.jboss.as.webservices.util.WSAttachmentKeys.JAXWS_ENDPOINTS_KEY;
 
 import java.util.Collections;
@@ -34,7 +33,6 @@ import org.jboss.as.server.deployment.DeploymentUnit;
 import org.jboss.as.server.deployment.annotation.CompositeIndex;
 import org.jboss.as.web.common.WarMetaData;
 import org.jboss.as.webservices.metadata.model.EJBEndpoint;
-import org.jboss.as.webservices.metadata.model.JAXRPCDeployment;
 import org.jboss.as.webservices.metadata.model.JAXWSDeployment;
 import org.jboss.as.webservices.metadata.model.POJOEndpoint;
 import org.jboss.as.webservices.webserviceref.WSReferences;
@@ -76,17 +74,6 @@ public final class ASHelper {
     }
 
     /**
-     * Gets list of JAXRPC EJBs meta data.
-     *
-     * @param unit deployment unit
-     * @return list of JAXRPC EJBs meta data
-     */
-    public static List<EJBEndpoint> getJaxrpcEjbs(final DeploymentUnit unit) {
-        final JAXRPCDeployment jaxrpcDeployment = getOptionalAttachment(unit, WSAttachmentKeys.JAXRPC_ENDPOINTS_KEY);
-        return jaxrpcDeployment != null ? jaxrpcDeployment.getEjbEndpoints() : Collections.<EJBEndpoint>emptyList();
-    }
-
-    /**
      * Gets list of JAXWS POJOs meta data.
      *
      * @param unit deployment unit
@@ -95,17 +82,6 @@ public final class ASHelper {
     public static List<POJOEndpoint> getJaxwsPojos(final DeploymentUnit unit) {
         final JAXWSDeployment jaxwsDeployment = unit.getAttachment(WSAttachmentKeys.JAXWS_ENDPOINTS_KEY);
         return jaxwsDeployment != null ? jaxwsDeployment.getPojoEndpoints() : Collections.<POJOEndpoint>emptyList();
-    }
-
-    /**
-     * Gets list of JAXRPC POJOs meta data.
-     *
-     * @param unit deployment unit
-     * @return list of JAXRPC POJOs meta data
-     */
-    public static List<POJOEndpoint> getJaxrpcPojos(final DeploymentUnit unit) {
-        final JAXRPCDeployment jaxrpcDeployment = unit.getAttachment(WSAttachmentKeys.JAXRPC_ENDPOINTS_KEY);
-        return jaxrpcDeployment != null ? jaxrpcDeployment.getPojoEndpoints() : Collections.<POJOEndpoint>emptyList();
     }
 
     /**
@@ -231,15 +207,6 @@ public final class ASHelper {
         if (wsDeployment == null) {
             wsDeployment = new JAXWSDeployment();
             unit.putAttachment(JAXWS_ENDPOINTS_KEY, wsDeployment);
-        }
-        return wsDeployment;
-    }
-
-    public static JAXRPCDeployment getJaxrpcDeployment(final DeploymentUnit unit) {
-        JAXRPCDeployment wsDeployment = unit.getAttachment(JAXRPC_ENDPOINTS_KEY);
-        if (wsDeployment == null) {
-            wsDeployment = new JAXRPCDeployment();
-            unit.putAttachment(JAXRPC_ENDPOINTS_KEY, wsDeployment);
         }
         return wsDeployment;
     }
